compiletest: Extract libtest-specific executor code to a submodule
